requirements.txt
clarifai
streamlit
Pillow
requests

